Commit du 17 juin 2013 : dernier idl + projet semi complet (reste à sauvegarder les sites visités dans le smartphone)

+++ src/AssistanceTouristique/ServiceDeplacementOperations.java
@@ -10,6 +10,6 @@ public interface ServiceDeplacementOperations
     /**
      * Operation calculerDureeTrajet
      */
-    public int calculerDureeTrajet(AssistanceTouristique.Coordonnees coordGPS_A, AssistanceTouristique.Coordonnees coordGPS_B, String mode);
+    public float calculerDureeTrajet(AssistanceTouristique.Coordonnees coordGPS_A, AssistanceTouristique.Coordonnees coordGPS_B, String mode);
 
 }
